onrev-proxy/requirements.txt
functions-framework==3.5.0
google-auth==2.33.0
requests==2.32.3
flask==3.0.3
